Add GarageService servlets, complete GarageSettings and change GarageSlot Ids to Long

+++ CarService/src/main/java/org/example/service/OrderService.java
@@ -41,8 +41,8 @@ public class OrderService {
         }
     }
 
-    public void assignRepairer(Order order, int... ids) {
-        for (int id : ids) {
+    public void assignRepairer(Order order, Long... ids) {
+        for (Long id : ids) {
             if (repairerService.getById(id).getIsAvailable()) {
                 order.addRepair(repairerService.getById(id));
                 repairerService.getById(id).setIsAvailable(false);
@@ -50,7 +50,7 @@ public class OrderService {
         }
     }
 
-    public void assignGarageSlot(Order order, int garageId) {
+    public void assignGarageSlot(Order order, Long garageId) {
         if (garageService.getById(garageId).isAvailable()) {
             if (order.getGarageSlot() != null) {
                 order.getGarageSlot().setAvailable(true); // освобождает гараж, который уже был назначен этому заказу
@@ -93,7 +93,7 @@ public class OrderService {
                     return orders.stream().sorted(Comparator.comparing(Order::isInProgress))
                             .collect(Collectors.toList());
                 case REPAIRER:
-                    orders.sort(Comparator.comparingInt(o -> o.getRepairers().iterator().next().getId()));
+                    orders.sort(Comparator.comparingLong(o -> o.getRepairers().iterator().next().getId()));
                     return orders;
             }
             return orders;
